[FIX] account@saas~12.4: performance when large amount of moves created
In the observed case,
More than 30.000 draft/cancelled invoices are converted to moves,
and therefore 30.000 moves are being created.

`NOT IN` or `!= ALL` is really not efficient
when using a large number of ids

https://stackoverflow.com/a/17824797

In the above thread,
it's discussed to use an anti-join,
or to create a temporary table with the ids to exclude.

As an alternative, we just use the ID of the first created move
and update only the lines of the moves with a smaller ID than the
first created move.

In the observed case,
the query never ends after hours of processing.
With this new strategy, the query ends in a few seconds.

Request 49194
M140416306132

closes odoo/upgrade#1641

Signed-off-by: Christophe Simonis <chs@odoo.com>