[FIX] base@saas~12.5: force sms upgrade for phone sanitize
The module `sms` can be installed during the upgrade if you have `crm` and `iap` installed
before upgrade as in 13.0,
it's auto-installed when `['base', 'iap', 'mail', 'phone_validation']`
and `phone_validation` is added in the dependences of `crm`

https://github.com/odoo/odoo/blob/13.0/addons/sms/__manifest__.py#L14
https://github.com/odoo/odoo/blob/13.0/addons/sms/__manifest__.py#L37
https://github.com/odoo/odoo/blob/13.0/addons/crm/__manifest__.py#L23

State of the modules before upgrade:
```sql
SELECT name, state FROM ir_module_module WHERE name in('crm', 'iap', 'sms', 'phone_validation');
       name       |    state
------------------+-------------
 phone_validation | uninstalled
 iap              | installed
 crm              | installed
 sms              | uninstalled
```

As `sms` is installed during the upgrade,
and not just "upgraded", its upgrade scripts are not executed,
meaning we miss the optimization regarding the computation of the `phone_sanitized` field.

Due to auto-install detection arcane, this script should be set in `post`, after the modules
changes from `13.0` are done.

Request 49898
M1509293041532

closes odoo/upgrade#1622

Signed-off-by: Christophe Simonis <chs@odoo.com>