[FIX] stock@11.0: `KeyError` when stock move has no quant
When the picking state is not `done` (e.g. `cancel)
`move_lots` keys are filled with the move ids from the result of the request

```
                cr.execute("""SELECT m.id, q.lot_id, ROUND(SUM(q.quantity):: numeric,%s::int)
                            FROM stock_move m, stock_quant q
                            WHERE q.reservation_id = m.id AND m.id IN %s
                            GROUP BY m.id, q.lot_id""", (rounding, tuple(moves.ids),))
```

If a move has no quant, the implicit join `FROM stock_move m, stock_quant q WHERE q.reservation_id = m.id`
will have as effect the move id won't be part of the result.

On the other hand, the loop is based on the `moves` ids rather than the move ids from the query result.
So, this is possible to have a `KeyError` when a move has no quants.

```
  File "/src/odoo/11.0/odoo/addons/base/maintenance/migrations/stock/11.0.1.1/post-10-advancedmoveopassign.py", line 46, in migrate
    if move_lots[move.id].get(ml.lot_id.id):
KeyError: 2295629
```

Request 52582
M120914764-zcaevxmq

closes odoo/upgrade#1651

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>