[FIX] sms@saas~12.5: sanitize 400.000 phone numbers
The database of the observed case has about 400.000 numbers.
Fetching all the numbers in once (with `cr.fetchall`)
leads to a `MemoryError`.

Therefore, split the numbers to fetch in chunks
Despite the heavy count of phone numbers to sanitize,
it's quite quick, something like 5 minutes,
so I don't think there is a need to output
a progress status.

Request 49898
M1509293041532

closes odoo/upgrade#1655

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>