[IMP] sale_amazon@saas~13.5: authentication model change
A module `sale_amazon_authentication` was introduced in 13.0 as part of
a rather large bugfix which changed the authentication mechanism in 13.0
for the Amazon Connector.

PR odoo/enterprise#12506 merged this fix module in the base
`sale_amazon` module.

This upgrade script ensure these changes are correctly migrated for
customers who had or had not installed the fix module.

Task-2279418