[FIX] base@0.0.0: delay user passwords encryption
When there are multiple thousands of passwords to encrypt,
it takes hours, even days, to encrypt all passwords.

We had the case several years ago for our own production.
To handle the case, we just skipped the encryption during the upgrade
(At the time, we commented the `init` method in `res_users.py`)
and we encrypted the password after the upgrade,
with a cron encrypting the passwords by chunks.

It doesn't really matter if the passwords are not encrypted right away:
It's still possible to login despite the password is not encrypted.
Actually, it will even encrypt the password on login.

This revision aims to do the same thing but genericly,
automatically,
for databases having more than 10.000 passwords to encrypt.

For the current observed case,
there are 664928 passwords to encrypt.

Request 49898
M1509293041532

closes odoo/upgrade#1658

Signed-off-by: Christophe Simonis <chs@odoo.com>