[FIX] account@saas~13.4: Fix migration of payment without account.move.line (#1665)
Before saas-13.4, the user was free to customize the journal entry linked to a payment as he want.
Because the link between account.payment and the accounting entries are only through the payment_id field on account.move.line, a payment could lost any reference to its related account.move.
In that case, the migration could fail due to payment without move_id field set.