[FIX] social_twitter@saas~14.1: remove inconsistent social streams
In odoo/enterprise#12538, we add a constraint to avoid having a social
stream of type "twitter_follow" or "twitter_likes" without
"twitter_followed_account_id".

This field is require in the form, but a bug allowed to create this kind
of stream. So, we remove those potential records from the database.

Task-2302570

closes odoo/upgrade#1706

Signed-off-by: Christophe Simonis <chs@odoo.com>