[IMP] base@0.0.0: add mechanism to ignore some field/view suppresion
Change in schema or view shouldn't be merged in stable, but in some cases, exception
can occur for bug fix or post-freeze `passe-droit`.

Upgrade tests can have trouble managing such case, adding a new field in saas-13.5 (master - 1)
will lead to error in master once the new shema is generated. The master will think
that a field has been removed. This is the reason why the 'template' tests will check that
the new generated database can still be upgraded to master before merging to avoid to break
master.

In some cases, backporting a feature from master can be enough for a new field or a new view,
and usually field or view removal in 13.5 is not a problem to forwardport, but both adding
and removing fields/views creates a merge deathlock, where only solution is to temporary
disable upgrade tests or add useless upgrade scripts.

This (temporary, I hope) solution allows to ignore some fields/view removal warnings
(causing most of the red builds). By giving a formated env variable, it is possible to change
log level of warnings and avoid raises. This should only be active for a short time, waiting
for the forward port to master. Disabling only specific warning will allow to continue to
test upgrade for other branches without taking the risk to miss something or blocking merge
during forwardport.

closes odoo/upgrade#1711

Related: odoo/odoo#56657
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>