[FIX] base@12.0: add `account_ponto` module
The module `account_ponto` was added in 12.0 after the release of 12.0
(and therefore after the writing of the upgrade scripts)
odoo/enterprise@364d4174ee335d4fe162cd29d7198daa1149d57a

This module should be installed automatically when `account_online_sync`
is installed. Without this revision, it wasn't the case.

opw-2333777
Request 48295

closes odoo/upgrade#1748

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>