[FIX] matt: `--tests` now works
It previously worked by accident, if the symlink was made from a
previous run. Now that runs are made in a temporary working directory,
the source version was never aware of the upgrade path, and thus never
found the upgrade tests.

This commit also adds some logging and fixes mypy type issues.

closes odoo/upgrade#1774

Signed-off-by: Christophe Simonis <chs@odoo.com>