[FIX] base@saas~12.5: no limit for image resizing
Do not limit the size when resizing the images:
This is mostly to avoid users to upload too large images,
but in the case of an upgrade, it's too late, the image
is already in database.
And this shouldn't block the upgrade.

```
 File "/src/odoo/13.0/odoo/fields.py", line 2094, in _process_related
    return self._image_process(super()._process_related(value))
  File "/src/odoo/13.0/odoo/fields.py", line 2088, in _image_process
    verify_resolution=self.verify_resolution,
  File "/src/odoo/13.0/odoo/tools/image.py", line 274, in image_process
    image = ImageProcess(base64_source, verify_resolution)
  File "/src/odoo/13.0/odoo/tools/image.py", line 89, in __init__
    raise ValueError(_("Image size excessive, uploaded images must be
smaller than %s million pixels.") % str(IMAGE_MAX_RESOLUTION / 10e6))
```

Request 47281
Request 54049

closes odoo/upgrade#1776

Signed-off-by: Christophe Simonis <chs@odoo.com>