[FIX] account@saas~12.4: Fix migration of anglo saxon lines with taxes
The mapping between account.invoice.line and account.move.line becomes very complex with anglo saxon lines because the price_subtotal of the invoice line doesn't match the accounting balance of the related move line.
Then, the mapping falls to less precise conditions and then, some invoice lines are matched with tax lines due to a missing condition when creating the mapping.
Also, this commit is improving the testing coverage of purchase_stock since:
https://github.com/odoo/upgrade/commit/bac3b53ccd814220296b371d39e4f49d866a31af

closes odoo/upgrade#1781

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>