[FIX] account_asset@saas~12.3: depreciation line into account_move
Deal with possible grouped journal entries for multiple depreciation
lines by creating a new journal and splitting the journal entries in
that journal.
This way, we keep all the journal entries untouched, removing the risk
of breaking the accounting.
If the user manually amended the linked journal entries, a balancing
journal entry will be created in the new journal.

Linked to upgrade requests 46548, 51908, 53090 and 53706
closes #1734

closes odoo/upgrade#1788

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>