[FIX] account@saas~12.4: fix draft/cancel invoice with same total moves
If a draft/cancelled invoice has a `move_id`,
and the move has the same amount,
there are good chances the move is the right move for the invoice.

If the move is posted, then the invoice should have been posted in the
first place as well. And the invoice line <-> move line mapping will do
its job.

If the move is not posted, then the invoice can be left
as draft/cancelled, but can be removed from the inconsistent
invoices list. The invoice line <-> move line mapping will
do its job to merge the draft/cancelled invoice with the unposted move.

Request 43639
Request 46864
Request 46382
Request 49863
Request 47486
Request 53420
Request 43616
Request 53120

closes odoo/upgrade#1787

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>