[FIX] account@0.0.0: attempt to set tags on taxes when none is set
The database has no tags set at all on its taxes,
while they are in 10.0 where they should have the taxes set.

This results from an upgrade from 8.0 to 9.0 were
the tags were not added on the taxes.

To try to solve this automatically,
we try to assign the tags according to the tax templates,
matching on the tax type, name and description to match
the tax with its template.

It's done in a `0.0.0` script so this automatic fix
is attempted whatever the current version of the database,
as soon as there is no tags at all configured

The table existance check on `account_tax_account_tag`
is done to not apply this fix for Odoo >= saas~12.3,
in which the table no longer exist.
It has been replaced by repartition lines on the taxes.

Request 51588
MY-32358

closes odoo/upgrade#1797

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>