[FIX] hr_payroll@saas~12.4: leaves with no set amount of time
It can be the case that some databases have `hr.leave`s with the same
start and end dates. This causes a problem when upgrading from
`saas~12.3` to `13.0`, as they are converted to records of the model
`hr.work.entry`, which does not allow records with the same start and
end dates.

In commit 26e06ac5d98c118b46a870166b7bb4343a27c4db, this issue was
addressed by modifying the end date of these records by adding the value
in the `number_of_days` column to the start date.

However, there is also the case that there are records for which the start
and end dates are the same and `number_of_days = 0`. To avoid having to
delete these records, the end date is set to be 1 microsecond more than the
start date, and as with the other case, the customer is warned about
these records so they can fix any possible issues with them later.

Request ID: 53720

closes odoo/upgrade#1800

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>