[FIX] account@saas~12.4: allow draft/cancelled invoices with moves
5013eaeb0335a21a0ffb32437875e5bc7d50963e
aims to automatically solve draft/cancelled invoice with a zero move,
by removing the link between the invoice and the pointless/zero move
so it create a new draft/move from the invoice

b19aeaef76da058cf7510e9e04196f0e12f9a118
aims to automatically solve draft/cancelled invoice associated to
a move with the same exact total than the invoice,
by updating the state of the invoice, which
should at first sight not be considered as draft/cancelled
if it has a posted move, with the same amount in addition.
In most cases, the "cancelled" state simply meant it was an invoice
which has been refunded, for which a credit note is done.

This commits aims to solve the remaining cases:
Solve automatically draft/cancelled invoice with a move with
a different amount. For those, we simply break the link
so a new draft/cancelled move is created for these invoices,
and the former move associated to the invoice is left untouched,
so the accounting is let identical.

Request 47486
Request 43616
Request 46561
Request 47343
Request 44250
Request 53585
Request 53789
Request 48953

closes odoo/upgrade#1803

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>