[FIX] account_asset@saas~12.5: disable noupdate of action_account_revenue_form
The action windoww `action_account_revenue_form` from the `account_asset` module is marked as `noupdate` in some databases. In `saas~12.5` some fields in the `account.asset` model change, making the domain definition in this action window obsolete, which causes the upgrade tests to fail.

To fix this, we force this action window to have `noupdate = false`.

Request ID: 53755

closes odoo/upgrade#1815

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>