[IMP] l10n_be_hr_payroll@saas~14.1: update fields in payroll.alloc.paid.leave
In this wizard, the date_start and date_end fields have been replaced
by the year field to select only the year of the period to compute the
legal leaves of each employee for the next year.

This migration script updates the record (if there are) and removes both
fields to create the year field.

closes odoo/upgrade#1832

Pr: https://github.com/odoo/enterprise/pull/13953
Taskid: 2314223
Related: odoo/enterprise#13953
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>