[FIX] documents@0.0.0: match_uniq on documents.tag
It has been the case that some users have messed with the definitions of
different tags defined in the `documents` module. The tags they
point to are modified and some records are even removed.

This causes problems during the upgrade if a record in `ir_model_data`
is removed, but the `document.tag` it pointed to still exists, but is
now pointed to by another record, since when updating the module, a
duplicate will be created.

To avoid this problem, we use `_match_uniq` to merge duplicated records.

Request ID: 54771

closes odoo/upgrade#1822

Signed-off-by: Christophe Simonis <chs@odoo.com>