[FIX] account@0.0.0: `account_account_tag_account_tax_template_rel` not in 14.0
Odoo 14.0 still has a table `account_tax_account_tag`
but it no longer as a many2many with as table
`account_account_tag_account_tax_template_rel`

Taxes now uses repartition lines to define the tags.
If we want to apply a similar "smart" migration
of the taxes tags, the script will need to be adapted
to use the repartition lines from now on.

M20052216491001

closes odoo/upgrade#1825

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>