[FIX] base@12.0: superuser in groups that inherit portal/public group
Since 12.0, a user cannot be at the same time in the `users` group and the
`portal` or `public` groups. This is checked by the `_check_one_user_type`
method. The upgrade scripts handled this by removing the user from some
of the groups if they overlapped. In particular, the root user was
removed explicitly from the `portal` and `public` groups.

The script however did not handle the case in which the root user is
not directly part of the `portal` or `public` groups, but is indirectly
part of one of them by being part of a group that inherits them. This
does not happen with standard Odoo modules, but some customisations may
lead to this particular case.

This is the case with the database from request 54302, in which the root
user is part of a custom group that inherits the `portal` group, which
makes the check fail.

Request ID: 54302

closes odoo/upgrade#1826

Signed-off-by: Christophe Simonis <chs@odoo.com>