[IMP] account@saas-12.3 and above: introduce tax report conversion helper
Tax reports often change from one version to the other in localization modules. When it happens, if we do nothing, tax configurations are not changed and the tags set on move lines are not modified either. This helper provides some generic handling to "translate" the old report into the new one.