[IMP] base@0.0.0: partially automatic module removal
When a module is added in stable it may look like the module has been removed
from the next version point of view. Runbot will add some information in the
environment to indicate that a new module is not being removed but
forwardported, and so error related to this module should be ignored.
This scripts remove the module automatically if it is found in the environment
and not in the addons path.
An optionnal minimal version can be given in case an upgrade script already
manage this module in a previous version (module removed/merged then
reintroduced latter)

closes odoo/upgrade#1836

Signed-off-by: Christophe Simonis <chs@odoo.com>