[FIX] stock@saas~12.4: some inventory locations not being migrated
In the upgrade to `saas~12.4`, in the model `stock.inventory`, the
`Many2one` field `location_id` is replaced by the `Many2many` field
`location_ids`.

In the upgrade scripts, only locations from `stock.inventory` records
which have state `draft` or `confirm`. However, some clients complain
that the locations set for records in other states have disappeared.

To fix this, all locations are migrated regardless of the state of the
`stock.inventory` record.

Request ID: 53088

closes odoo/upgrade#1840

Opw: 2323479
Signed-off-by: Christophe Simonis <chs@odoo.com>