[FIX] account@saas~13.2: limit sequence numbers to 9 digits
The current regex relies on other characters than digits
to decide what is the sequence prefix and number.

But for some bank statement names,
the prefix ends by numbers only,
and then a part of the prefix is considered
as the number by the regex,
and then the number of digits is higher than 9 digits,
therefore requiring a big integer:
```
psycopg2.errors.NumericValueOutOfRange: value "020200327375204268401739" is out of range for type integer
```

Here is the kind of prefix I got so far:
 - CAMT05300003120815600004
   the prefix seems to be CAMT053000031208156
   based on the occurences of 0. Other example to illustrate:
   CAMT05300002845860800001
   CAMT05300003054390300018
   CAMT05300000223258700002
   CAMT05300001246710900016
   This comes from the CAMT.053 bank statement format
 - 0201905280936225120000001
   the prefix is 020190528093622512, a date with time up to milliseconds,
   it seems,
 - 020200000062
   the prefix is 02020, just a year

I guess we could try to do something smart,
with cases about most known cases (date prefixes, camt prefixes, ...)
but at the end as
- the `number` fields is set as an integer,
- there is currently no possibility with the orm to convert it to a bigint,
- we cannot change the field type in stable 14.0

I see no other way than forcing the number to be 10 chars at most,
and then put the rest in the prefix.
Even if we do cases for most known case, this will always be a requirement.

Besides, I do not know what are the implications to not be correctly
accurate about splitting the prefix and number from the bank statement name,
if this is important for the reconcilation process or something.

M140526429516
M1509152980616
M1510013056373
M160425788835
M1606131095355
M1610031867766
M20020714499494

closes odoo/upgrade#1842

Signed-off-by: Christophe Simonis <chs@odoo.com>