[FIX] account@saas~12.4: prevent to match the counter-part line
In the invoice line <-> move line matching,
when there is only one line left to match in an invoice,
and the invoice line has in reality no matching line in the move,
such as a 0.00 amount invoice line,
for which there is no equivalent line in the move,
prevent to match that last line with the counter-part line.

In addition the matching to be wrong,
it completely fucks up the subtotal of the invoice,
as it match the invoice line with a move using an opposite value
(e.g. for a customer invoice, it matches a credit line with a debit line),
and therefore subtracts a part of the subtotal.

The example of the database subject of the opw below:
The invoice line 48624, being the last line of the invoice to match,
with a 0.00 amount, has been matched with the move line 83555,
the counter-part line of the move:

```sql
SELECT l.price_subtotal, a.code, a.name, a.internal_type
  FROM account_invoice_line l
  JOIN account_account a ON a.id = l.account_id
 WHERE l.id = 48624;

 price_subtotal |  code  |           name            | internal_type
----------------+--------+---------------------------+---------------
           0.00 | 701100 | Produits finis Maquillage | other

SELECT l.debit, l.credit, a.code, a.name, a.internal_type
  FROM account_move_line l
  JOIN account_account a ON a.id = l.account_id
 WHERE l.id = 83555;

  debit  | credit |  code  |                          name                           | internal_type
---------+--------+--------+---------------------------------------------------------+---------------
 1005.22 |    0.0 | 411100 | Clients - Ventes de biens ou de prestations de services | receivable
```

opw-2357352
M1704043632778

closes odoo/upgrade#1853

Signed-off-by: Christophe Simonis <chs@odoo.com>