[FIX] base@12.0.1.3: only one user type group
When migration "base/12.0.1.3/post-only-one-user-type-group.py" is executed,
it will delete all groups linked to the public user except the public group.

The group_show_line_subtotals_tax_included or group_show_line_subtotals_tax_excluded groups will also be unlinked from the public user.

If in the website settings, the option "Tax exclue" is checked,
prices will be displayed without tax for connected users (B2B view)
but will be displayed with tax for public users (B2C view).

There is a functional solution, it was enough to activate and deactivate the option "tax exclue" in the settings to remove the taxes from the price of public users.
From a practical point of view, this solution added the group_show_line_subtotals_tax_excluded group to the public user.

This fix corrects the following problem :
https://www.odoo.com/web#cids=1&id=2312351&menu_id=4720&model=project.task&view_type=form

Contract number : M19090212145595

closes odoo/upgrade#1854

Signed-off-by: Christophe Simonis <chs@odoo.com>