[IMP] stock@SaaS~14.1: add move reservation improvements
In order to provide more flexibility in when moves are auto-reserving
stock, 2 new fields are added to stock.picking.type:

1. 'reservation_method' = allows user to choose whether or not to
   auto-reserve and how it is done
2. 'reservation_days_before' = allows user to choose a day value before
   the move's (scheduled) date to do auto-reservation when appropriate
   method is selected

To support the 2nd value, the new field "reservation_date" is added to
stock.move in order to prevent having to recalcuate the appropriate date
every time the reservation_method applies to a move and needs to
auto-reserve.

Because the default reservation method matches the current behavior,
none of these added fields need to be computed for the upgrade.

Task: 2359317
COM PR: odoo/odoo#60300
ENT PR: odoo/enterprise#15050