[REM] procurement_jit@saas~14.1: replace procurement jit logic
Logic handled by the procurement jit has been moved into sale_stock now
that stock has new a stock.picking.type field that allows users do an
equivalent activation/deactivation of this logic.

We assume that if the module is uninstalled when sale_stock is installed
then that means the user manually uninstalled it (since it is auto-
installed when sale_stock is installed). Therefore in this case we
activate the equivalent 'manual' setting on all stock.picking.type
records that are outgoing (i.e. reproduce the deactivation of the
module).

Task: 2359317
COM PR: odoo/odoo#60300
Related ENT PR: odoo/enterprise#15050

closes odoo/upgrade#1868

Signed-off-by: Christophe Simonis <chs@odoo.com>