[FIX] helpdesk@saas~12.5: Recompute deadline before exceeded_days
In the upgrade to `saas~12.5`, a recompute of the field `exceeded_days` of
the model `helpdesk.sla.status` is executed. This model did not exist in
`saas~12.3`. The compute of this field uses another field, `deadline`,
which is also computed, but if it has not been computed before, it may
contain null values, causing the computation of `exceeded_days` to fail.

To fix this, the `deadline` field is also forced to recompute, before computing
`exceeded_days`.

Request ID: 54189

closes odoo/upgrade#1865

Signed-off-by: Christophe Simonis <chs@odoo.com>