[FIX] account@saas~12.4: ignore draft/cancelled invoices we can't create
There isn't much more cases of invalid invoices we do not
try to fix automatically, and for the rest of them
it becomes hard to find an automatic solution.

Instead, as those are draft and cancelled invoices,
we just skip their creation,
but still notify the users we did so in the migration
report.

Request 55598
M1711215752228

closes odoo/upgrade#1866

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>