[ADD] coupon@14.1: move mail_template_sale_coupon back to sale_coupon
pos_coupon module is added and during its development, it was found that
'mail_template_sale_coupon' only works for sale_coupon module because
of its dependence to a 'sale.order' record. In this regard, we are
returning this template back to sale_coupon. In pos_coupon module, we
create a similar template but is compatible with pos.order.

closes odoo/upgrade#1869

Community: https://github.com/odoo/odoo/pull/40325
Related: odoo/odoo#40325
Signed-off-by: Christophe Simonis <chs@odoo.com>