[FIX] account@saas~12.4: include grouped move lines in invoice lines
In Odoo <=12.0
If the boolean `group_invoice_lines` is set to True on the journal,
invoices are grouped by `tax_ids, tax_line_id, product_id,
analytic_account_id, date_maturity, analytic_tag_ids`
when creating the move for the invoice.

See
https://github.com/odoo/odoo/blob/d3b9f5c4e0cc67736e25008ec6d8c3b4324384da/addons/account/models/account_invoice.py#L1227-L1240

It's nothing new for upgrade: we were aware of this feature,
and convert invoice lines that have been grouped into one move line
to notes when merging invoices and moves together.

The move line grouping the invoice lines is left as is
in the move, and is not mapped to any invoice lines,
nor it's displayed in the invoice lines
(`exclude_from_invoice_tab` is left to True)

The problem raised by the customer is the fact the invoice subtotal
(`untaxed_amount`), is then wrong: it misses the grouped lines
subtotals.

Indeed, in Odoo 13.0, `untaxed_amount` is computed based on the
`balance` field of the lines included in the invoice lines
(`exclude_from_invoice_tab` to `False`)
https://github.com/odoo/odoo/blob/34527639873854d2be5a9bb1e3b5c8b88e52de33/addons/account/models/account_move.py#L1091-L1093

If we absolutely want `untaxed_amount` to be correctly computed,
I see only two solutions:
 - Set the balance in the note lines: The major drawback is that
   reports and other methods processing move lines might not filter out note lines,
   and therefore take into account the balance of these note lines,
   and therefore the numbers/totals will be wrong if thats the case.
   For instance, there is a constraint preventing to set debit/credit
   on a note line, but not balance.
   https://github.com/odoo/odoo/blob/34527639873854d2be5a9bb1e3b5c8b88e52de33/addons/account/models/account_move.py#L2665-L2667
   so this option seems to dangerous in my opinion
 - Include the move line grouping the invoice lines in the invoice lines
   (set `exclude_from_invoice_tab` to False)
   The major drawback is that there will be additional lines in the
   invoice lines compared to before (the note lines + the line that
   grouped them)

This revision aims to implement option 2.

Request 53318