[IMP] account@saas~14.1: make journal codes unique per company
This is done so that the new form of the journal code constraint (introduced by https://github.com/odoo/odoo/pull/59791 ) can be added correctly on every database.

Journal codes are not as critical today as they used to be. They only serve to identify the journals with a short string in the UI, and create the name suggestion for the first move made in them. The sequence of the moves now directly depend on the name of the previous move, and can be modified only by changing the name of a move. So, they are now totally disjoint from journal codes, which means we can modify them without auto-resequencing anything of what is in the journal.

closes odoo/upgrade#1880

Related: odoo/enterprise#14447
Related: odoo/odoo#59791
Signed-off-by: Christophe Simonis <chs@odoo.com>