[FIX] account@13.0: update mail aliases linked to journals
When upgrading to `13.0`, several customers have complained that their
mail aliases associated to accounting journals no longer work. This is
because in `13.0`, the `alias_defaults` needs to have the `journal_id`
of the journal linked to the alias.

Until now, the support team has been telling customers to manually
modify these defaults to fix this issue when it has been arisen.
However, it is something that should be taken care of during the
upgrade.

To fix this, aliases linked to journals are updated at the end of the
upgrade to `13.0`.

closes odoo/upgrade#1886

Opw: 2368825
Signed-off-by: Christophe Simonis <chs@odoo.com>