[FIX] account@saas~12.4: `only_one_line` match ignores `price_subtotal`
The `only_one_line` criteria is meant to be used,
with other criterias (e.g. `same_account`, `same_analytic_account`)
when there is only one line left to match in the invoice and the other
criterias matches.

With the condition
`AND ol.price_subtotal = il.price_subtotal`
instead of `there is only one line left in the invoice`,
it was more
`there is only one line of that amount in the invoice`.

This caused troubles when there are multiple invoice lines
with different amount left to match: the invoice lines
were matched with move lines having almost nothing in common
despite the account used (which is most of the time the same for all
products)

For instance, let's say the below
Invoice:
 - Product A, Qty 10, Price 100, Account A
 - Product B, Qty 5, Price 200, Account A
 - Product B, Qty 1, Price 40, Account A
 - Product C, Qty 5, Price 500, Account A
 - Product C, Qty 1, Price 100, Account A

Move:
 - Product A, Qty 10, Debit 100, Account A
 - Product B, Qty 6, Debit 240, Account A
 - Product C, Qty 6, Debit 600, Account A

So far, we could find a match for the first invoice line,
but for the others, we couldn't, because there was no match
in the move lines (same product, same qty, same price, ...).
There is therefore still 4 lines to match, all with different subtotals.

When using the conditions `{'only_one_line', 'same_account', 'same_analytic_account'}`,
the condition `only_one_line` always returned True, because there was
only one invoice line with the same subtotal.
It was therefore matching any invoice line with any move line because of
this, and it could end up matching a Product B invoice line with a
Product C move line.

Despite doing a bad matching, it was mainly problematic
for the invoice analysis, as the product/quantity/value
was completely messed up because of this.

It would make sense to match an invoice line
with a move line when there is only one line of that subtotal
in both the invoice and the move, but this is not what this condition is
doing: it compares the subtotal of one invoice line with the other
invoice lines rather than comparing the subtotal of the invoice line
with the subtotal of the move line.
This condition `ol.price_subtotal = il.price_subtotal` just doesn't make sense to me.

Request 53318

closes odoo/upgrade#1871

Signed-off-by: Christophe Simonis <chs@odoo.com>