[FIX] stock_account@saas~12.4: do not join on `account_move`
This is an oversight in #1891

As for the second `INSERT INTO stock_valuation_layer` request
which add the layers for automated/FIFO products,
this is possible multiple `account.move` have the same
`stock_move_id`, hence it multiplies the quantity and
valuation by the number of `account.move` sharing
the same `stock_move_id`. It shouldn't.

I am not sure if this is important or not to correctly
set the link `account_move_id` in the layer
for products not being FIFO/automated.
It wasn't the case before the above PR anyway,
so I just remove it in the mean time.
The quantities / valuation being multiplied is a far worse
issue than the missing link.

Request 55554

closes odoo/upgrade#1900

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>