[FIX] base@14.0: foreign key in Many2many
Before odoo 14. Foreign keys constraints were not checked on many2many relationship tables.
Since version 14, we check that the relationship tables do indeed have a foreign keys constraint towards each model. If not, the constraint is added.
In some tables, values already present can violate the constraint that will be added.

The solution is to run `fixup_m2m` on each relation table. We will therefore delete the id's which have a reference to non-existent records.

closes odoo/upgrade#1940

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>