[FIX] survey@saas~12.2: null values in survey.page title
In standard Odoo, the `title` field of `survey.page` is a required
field. However, some databases have modified this field to allow null
values. This creates a problem during the upgrade to `saas~12.3`, where
the `survey.page` model is merged with the `survey.question` model, as
if the new model does not allow null titles as well, the upgrade will
crash.

To fix this, if we encounter a null title, we use the record's ID
instead as a placeholder.

Request ID: 56037

closes odoo/upgrade#1917

Signed-off-by: Christophe Simonis <chs@odoo.com>